requirements.txt
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
contourpy==1.3.2
cycler==0.12.1
filelock==3.18.0
fonttools==4.57.0
fsspec==2025.3.2
gdown==5.2.0
idna==3.10
Jinja2==3.1.6
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.1
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.4
omp4py==0.1
packaging==24.2
pillow==11.2.1
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
requests==2.32.3
six==1.17.0
soupsieve==2.6
sympy==1.13.1
torch==2.6.0
torchinfo==1.8.0
torchvision==0.21.0
tqdm==4.67.1
typing_extensions==4.13.2
urllib3==2.4.0
